Premium Analytics: register widget-modules entity in the init module
Move the one-time core-data widgetModule entity registration out of the
dashboard route's beforeLoad and into the page init module, so it runs
once at app boot instead of on every navigation/preload.

The jetpack/v4 REST namespace now has a single source of truth in the
data package (DASHBOARD_REST_NAMESPACE), shared by the dashboard route
hooks and the init-module registration.

Addresses retrofox's review note on #49572.
