+++ projects/packages/premium-analytics/changelog/add-pa-widget-modules-init
@@ -1,0 +1,4 @@
+Significance: patch
+Type: changed
+
+Move the widget-modules entity registration to the page init module so it runs once at boot instead of on every dashboard beforeLoad.
