+++ projects/packages/premium-analytics/packages/init/package.json
@@ -8,9 +8,12 @@
 	"wpScriptModuleExports": "./build-module/index.mjs",
 	"dependencies": {
 		"@automattic/jetpack-script-data": "workspace:*",
+		"@jetpack-premium-analytics/data": "workspace:*",
 		"@wordpress/api-fetch": "7.50.0",
 		"@wordpress/boot": "0.17.0",
+		"@wordpress/core-data": "7.50.0",
 		"@wordpress/data": "10.50.0",
+		"@wordpress/i18n": "^6.9.0",
 		"@wordpress/icons": "^15.0.0"
 	}
 }
