+++ projects/packages/premium-analytics/packages/init/src/index.ts
@@ -2,9 +2,12 @@
  * External dependencies
  */
 import { getScriptData } from '@automattic/jetpack-script-data';
+import { DASHBOARD_REST_NAMESPACE } from '@jetpack-premium-analytics/data';
 import apiFetch from '@wordpress/api-fetch';
 import { store as bootStore } from '@wordpress/boot';
-import { dispatch } from '@wordpress/data';
+import { store as coreStore } from '@wordpress/core-data';
+import { dispatch, select } from '@wordpress/data';
+import { __ } from '@wordpress/i18n';
 import { chartBar } from '@wordpress/icons';
 
 // apiFetch middleware registers onto a shared, process-wide chain. Guard so
@@ -36,11 +39,45 @@ function setupApiFetch(): void {
 }
 
 /**
+ * Register the widget-modules discovery entity so the dashboard stage's
+ * `getEntityRecords` read resolves and feeds the records to `useWidgetTypes`.
+ * Premium Analytics serves the records from its own namespace (see
+ * `src/widget-modules.php`), independent of core's `wp/v2` endpoint. Runs once
+ * at boot. Guarded for idempotency in case init() ever runs more than once
+ * (re-mount, HMR, a future second boot).
+ */
+function registerWidgetModulesEntity(): void {
+	const coreSelect = select( coreStore ) as unknown as {
+		getEntityConfig: ( kind: string, name: string ) => unknown;
+	};
+	if ( coreSelect.getEntityConfig( 'root', 'widgetModule' ) ) {
+		return;
+	}
+
+	const coreDispatch = dispatch( coreStore ) as unknown as {
+		addEntities: ( entities: object[] ) => void;
+	};
+	coreDispatch.addEntities( [
+		{
+			name: 'widgetModule',
+			kind: 'root',
+			key: 'name',
+			baseURL: `/${ DASHBOARD_REST_NAMESPACE }/widget-modules`,
+			plural: 'widgetModules',
+			label: __( 'Widget modules', 'jetpack-premium-analytics' ),
+			supportsPagination: false,
+		},
+	] );
+}
+
+/**
  * Initialize the Jetpack Analytics app.
  * Runs before routes render.
  */
 export async function init(): Promise< void > {
 	setupApiFetch();
+
+	registerWidgetModulesEntity();
 
 	dispatch( bootStore ).updateMenuItem( 'dashboard', {
 		icon: chartBar,
