+++ projects/packages/premium-analytics/routes/dashboard/route.ts
@@ -3,14 +3,7 @@
  */
 import { getScriptData } from '@automattic/jetpack-script-data';
 import { ensureCoreSettingsReady, normalizeReportParams } from '@jetpack-premium-analytics/data';
-import { store as coreStore } from '@wordpress/core-data';
-import { dispatch, select } from '@wordpress/data';
-import { __ } from '@wordpress/i18n';
 import { redirect } from '@wordpress/route';
-/**
- * Internal dependencies
- */
-import { DASHBOARD_REST_NAMESPACE } from './hooks/constants';
 
 type DashboardSearch = Record< string, string | undefined >;
 
@@ -30,16 +23,9 @@ type DashboardSearch = Record< string, string | undefined >;
  * (core `site` settings not loaded yet) and the seeded `to` boundary would
  * land on a different instant than a later Apply writes.
  *
- * Then register the widget-modules discovery entity before the stage renders,
- * so the stage's `getEntityRecords` read resolves and feeds the records to
- * `useWidgetTypes`. Premium Analytics serves the records from its own namespace
- * (see `src/widget-modules.php`), independent of core's `wp/v2` endpoint.
- * Guarded for idempotency: beforeLoad re-runs on every navigation and preload.
- *
- * That registration is one-time bootstrap setup that could move to the page's
- * `init` module (`packages/init`) now that `@wordpress/build` supports it —
- * registering once at boot instead of on every beforeLoad run. Left here for
- * now (idempotency-guarded); tracked as a follow-up.
+ * The widget-modules discovery entity that feeds `useWidgetTypes` is registered
+ * once at app boot by the page's `init` module (`packages/init`), so it does not
+ * need to be (re-)registered here on every beforeLoad run.
  */
 export const route = {
 	beforeLoad: async ( { search }: { search?: DashboardSearch } = {} ) => {
@@ -98,27 +84,5 @@ export const route = {
 				search: seeded as unknown as never,
 			} );
 		}
-
-		const coreSelect = select( coreStore ) as unknown as {
-			getEntityConfig: ( kind: string, name: string ) => unknown;
-		};
-		if ( coreSelect.getEntityConfig( 'root', 'widgetModule' ) ) {
-			return;
-		}
-
-		const coreDispatch = dispatch( coreStore ) as unknown as {
-			addEntities: ( entities: object[] ) => void;
-		};
-		coreDispatch.addEntities( [
-			{
-				name: 'widgetModule',
-				kind: 'root',
-				key: 'name',
-				baseURL: `/${ DASHBOARD_REST_NAMESPACE }/widget-modules`,
-				plural: 'widgetModules',
-				label: __( 'Widget modules', 'jetpack-premium-analytics' ),
-				supportsPagination: false,
-			},
-		] );
 	},
 };
